+++ src/components/menu/submenu-block/index.d.ts
@@ -1,2 +1,0 @@
-export { default } from "./submenu-block";
-export * from "./submenu-block";