+++ src/components/multi-action-button/multi-action-button.component.tsx
@@ -19,9 +19,6 @@ import {
   filterOutStyledSystemSpacingProps,
 } from "../../style/utils";
 import { defaultFocusableSelectors } from "../../__internal__/focus-trap/focus-trap-utils";
-import Logger from "../../__internal__/utils/logger";
-
-let deprecatedWarnTriggered = false;
 
 export interface MultiActionButtonProps
   extends Omit<SplitButtonProps, "buttonType"> {
@@ -34,7 +31,6 @@ export interface MultiActionButtonProps
 export const MultiActionButton = ({
   align = "left",
   disabled,
-  as,
   buttonType,
   size,
   children,
@@ -44,14 +40,6 @@ export const MultiActionButton = ({
   "data-role": dataRole,
   ...rest
 }: MultiActionButtonProps) => {
-  if (!deprecatedWarnTriggered && as) {
-    deprecatedWarnTriggered = true;
-    Logger.deprecate(
-      // eslint-disable-next-line max-len
-      "The `as` prop is deprecated and will soon be removed from the `MultiActionButton` component interface. You should use the `buttonType` prop to achieve the same styling. The following codemod is available to help with updating your code https://github.com/Sage/carbon-codemod/tree/master/transforms/rename-prop"
-    );
-  }
-
   const ref = useRef<HTMLDivElement>(null);
   const buttonRef = useRef<HTMLButtonElement>(null);
   const buttonContainer = useRef<HTMLDivElement>(null);
@@ -234,7 +222,7 @@ export const MultiActionButton = ({
     onFocus: focusMainButton,
     onBlur: blurMainButton,
     onKeyDown: handleMainButtonKeyDown,
-    buttonType: buttonType || as,
+    buttonType,
     size,
     subtext,
     ...(!disabled && { onMouseEnter: showButtons }),